refactor: extract commands from `extension.ts` to separate files
